pcce-hackathon/services/job_scraper/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
playwright>=1.44.0
httpx>=0.27.0
linkedin-jobs-scraper>=2.3.0
pydantic>=2.7.0
redis>=5.0.4
python-dotenv>=1.0.1
asyncio; python_version < "3.10"
transformers>=4.40.0
torch --index-url https://download.pytorch.org/whl/cpu
